+++ RELEASE-NOTES.md
@@ -42,6 +42,7 @@ The Axom project release numbers follow [Semantic Versioning](http://semver.org/
 - Inlet: Added a new `Writer` for generating JSON schemas which can be used by text editors
   for autocompletion
 - Inlet: SphinxWriter will now document the signature of function callbacks added to a schema
+- Added support for QFunctions to `MFEMSidreDataCollection`
 
 ### Changed
 - The Sidre Datastore no longer rewires Conduit's error handlers to SLIC by default.